unit tests running from Pkg.test and they pass
