+++ test/runtests.jl
@@ -5,4 +5,5 @@ using Test
     include("test_combinations.jl")
     include("test_coverage_set.jl")
     include("test_greedy_tuples.jl")
+    include("test_parameter_order.jl")
 end
